requirements.txt
psycopg2-binary==2.9.5
rich==12.6.0
setuptools==49.2.1
pyshorteners==1.0.1